Fix loading plugin during boot
Some of the apps might try to load the plugin really early during boot
while user is not unlocked. At that time PackageManager doesn't allow
us to query the plugin and therefore we fallback to the static
implementation. Adding `android:directBootAware="true"` to the manifest
of the plugin makes it queryable even when user is not unlocked.

Bug: 279676520
Test: N/A
Change-Id: I4ea27cd741052fdb25d3f60da0394471aa463c11