Update CarUiListItemAdapter to be canonical for oems
With the support of two different layouts of CarUiListItemAdapter, it
doesn't make sense to always log a warning when any of the touch
interceptors aren't an instance of SecureView. So we only log when the
relevant touch interceptors are being used.

Also, update the ViewStub inflation logic to inflate all three touch
interceptors instead of just two so that it makes more sense from an oem
standpoint. Our reference should have three ViewStubs at the top level
car_ui_list_item file.

Bug: 293638796
Test: Manual
Test: atest CarUILibUnitTests

Change-Id: Id98e47e34f3e20b1f3298b0d1559e5bc94ccf095