Fix activity context leak in car-media-common lib
`PlaybackViewModel` is (indirectly) holding a context reference it is
initialized with. Object instances are retained via static field, so if an
activity is used for the context, it is retained even when the activity is
destoryed which results in a leak.

Swapping activity with application context solves the issue.

Bug: 317324773
Test: manual - triger a few unhandled configuration changes and check memory
profiler for detected leaks

Change-Id: Id5854ef67c5de35a061abf6654284bb5a49425df