Make QueueItemAdapterObserver compatible with Headers
This CL also fixes a crash where an observer calls a HeaderAdapter
function, but the HeaderAdapter isn't set for certain applications, and
is therefore null.

Additionally, the AdapterDataObserver should be attached to the full
ConcatAdapter in the case that a Header is used. onBindViewHolder should
also correctly bind the HeaderViewHolder rather than throwing an
IllegalArgumentException.

Bug: 336370273
Test: Manual

Change-Id: Idcce22aab09b4571ecad88a55337e8215e430604