Remember last played media session after boot
Store the last played active media source so on boot MediaSessionHelper
can default to that instead of the last playable one. This increases
consitency so the same app can be reflected before and after boot.
However as a consequence MediaSources may now have neither a
BrowseService nor a MediaController and can be instantiated with just a
package name.

Bug: 339004380
Bug: 341131977
Test: manual and testDebugUnitTest and atest car-media-common-unit-tests
Change-Id: Iabd6a64b30f5f2999838c314c4b65871b6b0ba88
Merged-In: Iabd6a64b30f5f2999838c314c4b65871b6b0ba88