Fix album art sometimes dissapearing
Album art will sometimes disappear if the PlaybackCardController gets null metadata. The controller sets the image visibilities to GONE and the ImageBinder isn't informed that the album state has changed. When valid metadata comes again, the ImageBinder sees a request for the already existing "current" image and does nothing. Without the callback the album art is not set to VISIBLE again.

Bug: 354037432
Test: Manual with tubi
Flag: EXEMPT minor bug fix
Change-Id: Id7984f3d1bddf769b9ea6d7bf6f178e9eceec68d