Fix b/355064996: Gradle build broken in full checkout
This fixes an issue caused by running a Gradle build within a checkout that includes the Soong build system. Code was previously added to read the Platform Sdk Version  from an environment variable if Soong was detected to be in the checkout. However, if the user had not "lunched" a target then the variable would not be set and would return an empty string. This caused a build failure.

BUG: 355064996
Test: Check out the full branch without using a `-g` flag to filter the repositories, then run `./gradlew assembleDebug`
Change-Id: Ife98b5a9ab9dd308749618ca05bbc5fdb5aa2c6f