Reduce Gradle daemon heap size
Reduce the maximum heap size for the Gradle daemon from 8192m to 4096m and set the initial heap size to 2048m.

The heap was increased to 8192m as a quick fix for build failures, but there is no real benefit to allowing the heap to get that big. In addition, setting the initial heap size will speed up builds by allocating most of the memory for the build once during startup.

Bug: 356381551
Test: ./gradlew assemble assembleAndroidTest test

Change-Id: I6ab80ea04438a0244cc4a71fd2a95da81bf90c8b