Filter media sessions for media notifications
Phone apps have the assumption that the user will only be able to
control media session playback when backgrounded by using the media
notification. This change updates the MediaSessionHelper logic to filter
out media sessions that don't have an associated media notification.

This fixes potential issues such as the app not having the legal right
for background play, or when the system will kill the media session due
to it not being a foreground service.

Bug: 357865809
Test: Manual on VIC and SC emulators using media notification and
non-media notification apps.
Flag: EXEMPT bug_fix

Change-Id: I30ab45230576ee1215fbb7df94feb0c3f2cf0712