Configure Gradle to use AOSP's JDK 17
This commit configures Gradle to use the JDK from the AOSP checkout for the build process. It adds to the ./gradlew script to find and use the JDK in prebuilds/jdk. This code was cropped from AndroidX: https://source.corp.google.com/h/android/platform/superproject/+/refs/heads/androidx-main:frameworks/support/gradlew

This ensures that the correct JDK is used for
 building the project, especially on CI machines where `JAVA_HOME` might point to an older JDK.

Test: ./gradlew -q javaToolchains

Change-Id: Ib161cc6a46c284b181f444fc47bc22f698455dba