feat: Introduce project plugin
This commit introduces a new project plugin to manage common configurations for Android projects.

The plugin sets the `compileSdk` and `
buildToolsVersion` based on the versions defined in the `libs.versions.toml` catalog. It also applies the `aaosApps.buildLogic` plugin to all projects.

The new plugin needs to be in a seperate Gradle project, so the settings plugin's Gradle project has been renamed to "settingsPlugin"

Test: ./gradlew assemble assembleDebug test

Change-Id: Ic2de4bdf63136cce4ff2df90d707190de7936fb8