Update Dependencies for car-app-card-lib
This change is the first of an effort to update the dependencies of each Gradle module to current, up-to-date versions.

Androidx dependencies have been updated to their latest versions, roughly matching the current versions in main (though Androidx occasionally updates their prebuilts before the same version was released publicly).

Third party dependencies have been updated to the versions used for the Soong builds where possible. A notable exception is mockito, which seems to have been abandoned and is five years and counting out of date.

Test: manual

Change-Id: I363806218c8cdfd2d9e78a1260d6dd4fb009b2e7