Revert "Change the Gradle build to use pre-fetched Gradle binary."
This reverts commit d102576298a53199f2985bcea2ad73804e49cc60.

Reason for revert: I've realized that Android Studio is re adding the original gradle wrapper script and jar when I run a sync. I'd thought I was dealing with the files being left behind when I switched branches but it turns out that's not the case. I'll have to figure out how to prevent that.

Change-Id: I764e9a004cfcc8520146de98f3e52c8f5f0d6007