Busytown: Add run-unit-tests.sh script and remove unit tests from build-and-test.sh
This change introduces a new script, run-unit-tests.sh, dedicated to running Busytown's unit tests.
The script sets up the build environment and then executes the tests using Gradle.

This does mean that the "build-and-test.sh" script's name isn't very accurate any more, but changing that would require changing the busytown configs (which is a lot of work to get right)

To streamline the main build process, the unit test execution has been removed from build-and-test.sh. This allows for greater flexibility and separation of concerns in the build pipeline.

Test: ./busytown/build-and-test.sh; ./busytown/run-unit-tests.sh

Bug: b/391863828
Change-Id: I4adc5c97430b8f1e3269a0736a3fae06119dc904