Fix Forward "Update NDK version to 26.2.11394342 and set NDK ..."
The previous CL was reverted because it ran afoul of
`androidmk_denylist.go`. I've checked it and the `//toolchain` directory is still allowed, and that's where the NDK is actually supposed to live. So I've changed the manifest to place the NDK where it should be in toolchain and ran additional abtd tests to try to catch any other issues.

DISABLE_TOPIC_PROTECTOR - doesn't work with manifest changes

Test: ./gradlew build and presubmits

Change-Id: I1667af06ec56bdc3fc2e7c49a774d6e57648d29c