Add full NDK support, update to r27
This change adds full NDK support to builds using the prebuilt SDK
This change is a bit tricky because there's the NDK, CMake and Ninja.
You can set the NDK dir in the Gradle config, but you have to set the
cmake dir in local.properties. The layout of these tools in prebuilts are different
than how they should be in a normal Android SDK install, meaning we need
to set the locations manually. I made a quick attempt to get the
prebuilts laid out the same way as a normal Android SDK, but it wasn't
going well so I just set it up the way AndroidX does it.

Note: For anyone who insists on trying to use the prebuilt SDK for
day-to-day development, rather than the regular SDK, (I don't recommend
this) you're going to
have to set your `cmake.dir` in your local.properties manually.

Test: ./gradlew assembleDebug and ./busytown/build-and-test.sh on both
Mac and Linux

Change-Id: I69e23633f9bebbf78e53fe96d0baacb3e8b50c01