Configure Gradle toolchain provisioning
This enables a new feature in Gradle 8.13, allowing Gradle to specify the JDK to be used for the Gradle daemon. This will allow both Android Studio and CLI builds to use the same JDK.

In addition, I'm adding Gradle's Java toolchain fetcher plugin. With this, if a JDK is not available on your machine, Gradle will be able to go out and download the JDK for you. I've left the JDK vendor default, which will resolve to Adoptium.

The prebuilt JDKs will still be used for BuildBot builds. Neither myself nor the Android Gradle team are expecting any issues from this.

Test: ./busytown/smoke-test.sh
Change-Id: I1c095345f56190e763280a05cab40560483423ca