Refactor code for finding repo root and setting build dir
This change consolidates the logic for finding the root of the repository and setting the build directories for all projects.

Logic for finding the root has been moved to a "Script Plugin", which is generally discouraged. However, since this code is used to move the build directories there isn't a way to write it as a precompiled plugin (because it would then create a build directory)

While working on this I realized that there wasn't any point in moving the build directory from within the context of Settings, because from within the Settings context I was calling a Gradle hook to move the build directory after the Project had been intialized anyways. That was the only thing the Settings plugin was doing, so I just deleted it. I had over-engineered it by trying to do it the "right" way.

Test: ./busytown/build-and-test
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:a92bf1c106e3fc7171123c57a555451a7221a214)
Merged-In: Ia554f136a09aaed4e2df48c76e50b5a0803a23dd
Change-Id: Ia554f136a09aaed4e2df48c76e50b5a0803a23dd