Update car-testing-common to use version catalog and remove redundant test rule
Migrates from the deprecated InstantTaskExecutorRule in car-testing-common to the androidx equivalent. This ensures compatibility with the latest architecture components and aligns with current best practices.

The file is identical to the source for the same rule in androidx.arch.core.testing, so there's really no reason to keep it. It relied on some internal AndroidX APIs and wouldn't compile with up to date dependencies.

Test: ./busytown/build-and-test and ./busytown/run-host-unit-tests
Change-Id: I6fc3597e601f5e56c3fcb42745b00103d786ca95